Add SyncCommitteeService

+++ consensus/types/src/sync_committee_subscription.rs
@@ -9,7 +9,7 @@ pub struct SyncCommitteeSubscription {
     /// The validators index.
     pub validator_index: u64,
     /// The sync committee indices.
-    pub sync_committee_indices: Vec<usize>,
+    pub sync_committee_indices: Vec<u64>,
     /// Epoch until which this subscription is required.
     pub until_epoch: Epoch,
 }
